Streamline Boost tests for compiled libraries and add documentation on Boost module usage
