+++ test/boost_libs/cmake/autocmake.cfg
@@ -5,8 +5,17 @@ min_cmake_version: 2.8
 [cxx]
 source: ../../../modules/cxx.cmake
 
+[mpi]
+source: ../../../modules/mpi.cmake
+
+[python_interpreter]
+source: ../../../modules/python_interpreter.cmake
+
+[python_libs]
+source: ../../../modules/python_libs.cmake
+
 [boost]
-override: {'major': 1, 'minor': 59, 'patch': 0, 'components': 'chrono;timer;system'}
+override: {'major': 1, 'minor': 59, 'patch': 0, 'components': 'mpi;serialization;python'}
 source: ../../../modules/boost/boost.cmake
 
 [default_build_paths]
