--- test/boost_mpi_libs/cmake/autocmake.cfg
@@ -1,19 +1,0 @@
-[project]
-name: example
-min_cmake_version: 2.8
-
-[cxx]
-source: ../../../modules/cxx.cmake
-
-[mpi]
-source: ../../../modules/mpi.cmake
-
-[boost]
-override: {'major': 1, 'minor': 59, 'patch': 0, 'components': 'mpi;serialization'}
-source: ../../../modules/boost/boost.cmake
-
-[default_build_paths]
-source: ../../../modules/default_build_paths.cmake
-
-[src]
-source: ../../../modules/src.cmake
