--- test/boost_python_libs/cmake/autocmake.cfg
@@ -1,22 +1,0 @@
-[project]
-name: example
-min_cmake_version: 2.8
-
-[cxx]
-source: ../../../modules/cxx.cmake
-
-[python_interpreter]
-source: ../../../modules/python_interpreter.cmake
-
-[python_libs]
-source: ../../../modules/python_libs.cmake
-
-[boost]
-override: {'major': 1, 'minor': 56, 'patch': 0, 'components': 'python'}
-source: ../../../modules/boost/boost.cmake
-
-[default_build_paths]
-source: ../../../modules/default_build_paths.cmake
-
-[src]
-source: ../../../modules/src.cmake
