--- test/boost_python_libs/src/CMakeLists.txt
@@ -1,7 +1,0 @@
-add_executable(example example.cpp)
-
-if(BUILD_CUSTOM_BOOST)
-    add_dependencies(example custom_boost)
-endif()
-
-target_link_libraries(example ${Boost_PYTHON_LIBRARY} ${PYTHON_LIBRARIES})
